% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamline-to-animate.R
\name{animation_transition}
\alias{animation_transition}
\title{Transition a streamplot to an animation ready object}
\usage{
animation_transition(plot)
}
\arguments{
\item{plot}{A ggplot object from which the data and other plot attributes
will be extracted to create the animation transition.}
}
\value{
A ggplot object with an animation transition effect applied.
}
\description{
\code{\link[=animation_transition]{animation_transition()}} takes a streamline plot and alters it so that it
will work with gganimate syntax.
}
\examples{
\dontrun{

library(gganimate)

# Example user-defined function
f <- function(v) {
  x <- v[1]
  y <- v[2]
  c(-1 - x^2 + y, 1 + x - y^2)
}

# Create a ggplot with the stream plot layer
p <- ggplot() +
  geom_streamplot(
    aes(rownum = after_stat(rownum)),
    fun = f, xlim = c(-3, 3), ylim = c(-3, 3)
  ) +
  coord_fixed() +
  theme_bw()

# Create an animation transition plot
anim <- animation_transition(plot = p) +
  transition_reveal(rownum) +
  ease_aes('linear')

# Animate the plot
anim <- animate(anim, nframes = 25, fps = 20, end_pause = 0, renderer = gifski_renderer())
}

}
