% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efield.R
\name{efield}
\alias{efield}
\alias{efield_maker}
\title{Electric field}
\usage{
efield(u, charge_positions, charges, k = 1, q_test = +1)

efield_maker(charge_positions, charges, k = 1, q_test = +1)
}
\arguments{
\item{u}{The position of the test charge.}

\item{charge_positions}{The positions of the fixed charges generating the
electric field.}

\item{charges}{The charges of the points placed at the positions of
charge_positions.}

\item{k}{The constant of proportionality, defaulted to 1. See examples for a
more rigorous use of physical constants.}

\item{q_test}{The test charge, defaulted to +1.}
}
\value{
A vector containing the force felt by the test charge on account of
the electric field.
}
\description{
The vector field generated by collection of fixed electrical charges, as
dictated by Coulomb's law. This function is mainly used to provide examples
for visualizing vector fields with ggvfields.
}
\examples{

# set a - charge at (-1,-1) and a + charge at (1,1)
charge_positions <- rbind(c(-1,-1), c(1,1))
charges <- c(-1, +1)


# calculate force on test charge (+1) at c(0,1), ignoring physical constants
efield(c(0,1), charge_positions, charges)


# visualize electric field
ggplot() +
  geom_vector_field(
    fun = function(u) efield(u, charge_positions, charges),
    xlim = c(-2,2), ylim = c(-2,2)
  ) +
  scale_color_viridis_c(trans = "log10")


# visualize electric field using efield_maker, which is a simpler wrapper
ggplot() +
  geom_vector_field(
    fun = efield_maker(charge_positions, charges),
    xlim = c(-2,2), ylim = c(-2,2)
  ) +
  scale_color_viridis_c(trans = "log10")


# electric constant from https://en.wikipedia.org/wiki/Vacuum_permittivity
ep0 <- 8.854187818814e-12
k <- (4*pi*ep0)^-1
efield(c(0,1), charge_positions, charges, k)

}
\seealso{
\url{https://en.wikipedia.org/wiki/Coulomb\%27s_law}
}
