% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{ensure_nonempty_data}
\alias{ensure_nonempty_data}
\title{Ensure Non-empty Data}
\usage{
ensure_nonempty_data(data)
}
\arguments{
\item{data}{The input data to check.}
}
\value{
A data frame. If the input data is empty, returns a default data frame with a single row.
}
\description{
This function ensures that the data is non-empty. If the input data is empty, it returns a default data frame.  ggplot2 layers require a data argument but this allows layers to operate without data.
}
\keyword{internal}
