% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hexagonal_lattice.R
\name{generate_hexagonal_lattice}
\alias{generate_hexagonal_lattice}
\title{Generate a Hexagonal Lattice}
\usage{
generate_hexagonal_lattice(xlim, ylim, d)
}
\arguments{
\item{xlim}{A numeric vector of length 2 specifying the x-axis limits.}

\item{ylim}{A numeric vector of length 2 specifying the y-axis limits.}

\item{d}{A numeric value specifying the hexagon diameter.}
}
\value{
A data frame with two columns, \code{x} and \code{y}, containing the coordinates of the hexagonal grid points.
}
\description{
This function generates a hexagonal lattice of points within the given x and y limits,
using a specified hexagon diameter.
}
\examples{
xlim <- c(-10, 10)
ylim <- c(-10, 10)
d <- 0.5
hex_lattice <- generate_hexagonal_lattice(xlim, ylim, d)
}
