% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_complex_function.R
\name{geom_complex_function}
\alias{geom_complex_function}
\title{Create a Complex Vector Field Geom Layer}
\arguments{
\item{fun}{A user-defined function that takes a complex number as input and
returns a complex number.}

\item{relim}{A numeric vector of length 2 giving the real-axis limits.}

\item{imlim}{A numeric vector of length 2 giving the imaginary-axis limits.}

\item{n}{An integer specifying the number of grid points along each axis.}
}
\value{
A ggplot2 layer that can be added to a ggplot object to produce a
vector field plot.
}
\description{
\code{\link[=geom_complex_function]{geom_complex_function()}} generates a raster plot layer of a user-defined
complex function. The hue of each color corresponds to the argument (angle)
of the complex number resulting from the function, indicating the direction
of the vector. The lightness of the color reflects the magnitude of the
vector, providing a visual representation of both direction and intensity.
}
\examples{
#library(ggvfields)
# Example user-defined function
f <- function(z) (z^2 + 1) / (z^2 - 1)

# Create a ggplot with the vector field layer
#ggplot() +
 # geom_complex_function(fun = f, relim = c(-2, 2), imlim = c(-2, 2), n = 100) +
 # labs(x = "Real", y = "Imaginary")
}
\keyword{internal}
