% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_function_2d.R
\docType{data}
\name{geom_function_2d}
\alias{geom_function_2d}
\alias{stat_function_2d}
\alias{StatFunction2d}
\alias{GeomFunction2d}
\title{Create a 2D Function Visualization Layer of the Norm of a Vector Field}
\format{
An object of class \code{StatFunction2d} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.

An object of class \code{GeomFunction2d} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
geom_function_2d(
  mapping = NULL,
  data = NULL,
  stat = StatFunction2d,
  geom = GeomFunction2d,
  ...,
  position = "identity",
  fun = NULL,
  xlim = NULL,
  ylim = NULL,
  n = NULL,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_function_2d(
  mapping = NULL,
  data = NULL,
  geom = GeomFunction2d,
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  fun = NULL,
  xlim = c(-1, 1),
  ylim = c(-1, 1),
  n = 50
)

StatFunction2d

GeomFunction2d
}
\arguments{
\item{mapping}{Aesthetic mappings, created using \code{aes()}. If \code{NULL}, defaults are used.}

\item{data}{Optional data frame to override the default data source.}

\item{stat}{Statistical transformation to use. Defaults to \code{StatFunction2d}.}

\item{geom}{Geom used for rendering. Defaults to \code{GeomFunction2d}.}

\item{...}{Other arguments passed to the layer, such as additional parameters.}

\item{position}{Position adjustment for the layer. Defaults to \code{"identity"}.}

\item{fun}{A function that takes a matrix of x, y values and returns a matrix of dx, dy values.}

\item{xlim}{Numeric vector of length 2 specifying the x-range of the grid. Required if \code{fun} is provided.}

\item{ylim}{Numeric vector of length 2 specifying the y-range of the grid. Required if \code{fun} is provided.}

\item{n}{Number of points in the grid along each axis. Defaults to \code{11} in \code{stat_function_2d}.}

\item{show.legend}{Logical. Should this layer be included in the legends? \code{NA} includes if aesthetics are mapped.}

\item{inherit.aes}{If \code{FALSE}, overrides default aesthetics rather than combining them.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to \code{FALSE}.}
}
\value{
A \code{ggplot2} layer.
}
\description{
\code{geom_function_2d} adds a layer to visualize 2D functions or vector fields in a \code{ggplot2} plot.
}
\examples{
# Function that calculates the norm
f <- function(v) {
  x <- v[1]
  y <- v[2]
  c(sqrt(x^2 + y^2))
}

ggplot() +
  geom_function_2d(fun = f, xlim = c(-5, 5), ylim = c(-5, 5))

# Sinusoidal combination of sine and cosine
f_sin_cos <- function(v) {
  x <- v[1]
  y <- v[2]
  sin(x) * cos(y)
}

ggplot() +
  geom_function_2d(fun = f_sin_cos, xlim = c(-5, 5), ylim = c(-5, 5))

# Gaussian bump function
f_gaussian <- function(v) {
  x <- v[1]
  y <- v[2]
  exp(-(x^2 + y^2) / 2)
}

ggplot() +
  geom_function_2d(fun = f_gaussian, xlim = c(-5, 5), ylim = c(-5, 5))

# Radial sine wave function
f_radial_wave <- function(v) {
  x <- v[1]
  y <- v[2]
  r <- sqrt(x^2 + y^2)
  sin(r)
}

# Some functions need more resolution for clarity
ggplot() +
  geom_function_2d(fun = f_radial_wave, xlim = c(-50, 50), ylim = c(-50, 50), n = 100)

# Complex combination of radial and angular components
f_complex <- function(v) {
  x <- v[1]
  y <- v[2]
  r <- sqrt(x^2 + y^2)
  theta <- atan2(y, x)
  sin(r) * cos(theta)
}

ggplot() +
  geom_function_2d(fun = f_complex, xlim = c(-50, 50), ylim = c(-50, 50), n = 500)

# Spiral pattern function
f_spiral <- function(v) {
  x <- v[1]
  y <- v[2]
  r <- sqrt(x^2 + y^2)
  theta <- atan2(y, x)
  sin(r + theta)
}

ggplot() +
  geom_function_2d(fun = f_spiral, xlim = c(-50, 50), ylim = c(-50, 50), n = 500)
}
\keyword{datasets}
