% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_streamline.R
\docType{data}
\name{geom_streamplot}
\alias{geom_streamplot}
\alias{GeomStreamplot}
\alias{stat_streamline}
\alias{StatStreamplot}
\title{Create a Stream Plot Geom Layer}
\usage{
geom_streamplot(
  mapping = NULL,
  data = NULL,
  stat = "streamplot",
  position = "identity",
  na.rm = FALSE,
  show.legend = TRUE,
  inherit.aes = TRUE,
  fun,
  xlim,
  ylim,
  max_length,
  max_steps,
  ds,
  min_dist,
  mask_shape_type = "square",
  arrow = grid::arrow(angle = 20, length = unit(0.015, "npc"), type = "closed"),
  chop = TRUE,
  chop_fraction = 0.1,
  ...
)

stat_streamline(
  mapping = NULL,
  data = NULL,
  stat = "streamplot",
  position = "identity",
  na.rm = FALSE,
  show.legend = TRUE,
  inherit.aes = TRUE,
  fun,
  xlim,
  ylim,
  max_length,
  max_steps,
  ds,
  min_dist,
  mask_shape_type = "square",
  arrow = grid::arrow(angle = 20, length = unit(0.015, "npc"), type = "closed"),
  chop = TRUE,
  chop_fraction = 0.1,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, either as a \code{ggproto} \code{Geom} subclass or as a string naming the
stat stripped of the \code{stat_} prefix (e.g. \code{"count"} rather than
\code{"stat_count"})}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{fun}{A user-defined function that takes a numeric vector as input and
returns a numeric vector of the same length.}

\item{xlim}{A numeric vector of length 2 giving the x-axis limits.}

\item{ylim}{A numeric vector of length 2 giving the y-axis limits.}

\item{max_length}{A numeric value specifying the maximum length of
streamlines.}

\item{max_steps}{An integer specifying the maximum number of steps for
streamline generation.}

\item{ds}{A numeric value specifying the distance between steps for
streamline generation.}

\item{min_dist}{A numeric value specifying the minimum distance between
streamlines.}

\item{mask_shape_type}{A character string specifying the mask shape type:
"square", "diamond", or "inset_square".}

\item{arrow}{Arrow specification, as created by \code{grid::arrow()}.}

\item{chop}{A logical value indicating whether to chop the trajectories into
segments.}

\item{chop_fraction}{A numeric value specifying the maximum allowable segment
size as a fraction of the total range for chopping the trajectories.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\value{
A ggplot2 layer that can be added to a ggplot object to produce a
stream plot.
}
\description{
\code{\link[=geom_streamplot]{geom_streamplot()}} generates a stream plot layer of a user-defined vector
field function. The lines in the plot represent the flow of data points
through the vector field.
}
\examples{
library(ggplot2)
# Example user-defined function
f <- function(v) {
  x <- v[1]
  y <- v[2]
  c(-1 - x^2 + y, 1 + x - y^2)
}

# Create a ggplot with the stream plot layer
ggplot() +
  geom_streamplot(
    aes(group = after_stat(id)), fun = f, xlim = c(-3, 3),
        ylim = c(-3, 3), max_length = 10000, max_steps = 10000,
        ds = .05, min_dist = .25, chop = TRUE, chop_fraction = 0.1,
        mask_shape_type = "diamond"
        ) +
  coord_fixed() +
  theme_minimal()

}
\keyword{datasets}
