% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_vector.R
\docType{data}
\name{geom_vector}
\alias{geom_vector}
\alias{stat_vector}
\alias{StatVector}
\title{geom_vector}
\format{
An object of class \code{StatVector} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
geom_vector(
  mapping = NULL,
  data = NULL,
  stat = StatVector,
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  arrow = grid::arrow(angle = 25, length = unit(0.025, "npc"), type = "closed"),
  center = FALSE
)

stat_vector(
  mapping = NULL,
  data = NULL,
  geom = GeomStream,
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  arrow = grid::arrow(angle = 25, length = unit(0.025, "npc"), type = "closed"),
  center = FALSE
)

StatVector
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{ggplot2::aes()}}. \strong{Required:}
\code{x} and \code{y} must always be provided. In addition, either \code{xend} and \code{yend} or
\code{angle} and \code{distance} must be supplied. No grouping (or id) is needed.}

\item{data}{A data frame containing the vector data in wide format.}

\item{stat}{The statistical transformation to use on the data for this layer.
Defaults to \link{StatVector}.}

\item{position}{Position adjustment, either as a string or the result of a call
to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{ggplot2::layer()}}.}

\item{na.rm}{Logical. If \code{FALSE} (the default), missing values are removed with a warning.}

\item{show.legend}{Logical. Should this layer be included in the legends?}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics rather than combining with them.}

\item{arrow}{An optional \code{\link[grid:arrow]{grid::arrow()}} specification to add arrowheads to the vectors
(default: \code{grid::arrow(angle = 25, length = unit(0.025, "npc"), type = "closed")}).}

\item{center}{Logical. If \code{TRUE}, the vector is recentered so that the original
(x, y) becomes the midpoint of the vector. Default is \code{FALSE}.}

\item{geom}{The geometric object used to draw the streamline. Defaults to
\link[ggplot2:ggplot2-ggproto]{ggplot2::GeomPath} in \code{geom_stream()}, or \link{GeomStream} in \code{stat_stream()}.}
}
\value{
A ggplot2 layer that can be added to a plot.
}
\description{
A convenient ggplot2 layer for drawing vectors. This function accepts
wide-format data with the aesthetics \code{x} and \code{y} plus either (\code{xend}, \code{yend})
or (\code{angle}, \code{distance}). If the latter is supplied, the endpoints are computed
as a translation of the starting point using polar coordinates (assuming
the angle is in degrees). The data is then converted into long format (two rows
per vector) using \link{StatVector} and plotted using \link{GeomStream}. Directional
arrowheads can be added to indicate the vector direction.
}
\examples{
vectors1 <- data.frame(
  x    = c(0, 1, 2),
  y    = c(0, 1, 2),
  xend = c(3, 1, 5),
  yend = c(0, 5, 6)
)
ggplot(vectors1, aes(x = x, y = y, xend = xend, yend = yend)) +
  geom_vector(center = TRUE)

vectors2 <- data.frame(
  x        = c(0, 1, 2),
  y        = c(0, 1, 2),
  angle    = c(0, 90, 45),
  distance = c(3, 4, 5)
)
ggplot(vectors2, aes(x = x, y = y, angle = angle, distance = distance)) +
  geom_vector(center = TRUE)

}
\keyword{datasets}
