% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvfields-package.R
\name{ggvfields}
\alias{ggvfields}
\alias{ggvfields-package}
\title{ggvfields: Vector Field Visualizations with ggplot2}
\description{
A ggplot2 extension for visualizing vector fields in a two-dimensional space.
Provides functions to create vector field layers using user-defined vector
field functions.
}
\seealso{
Useful links:
\itemize{
\item \url{https://jamesotto852.github.io/ggdensity/}
\item \url{https://github.com/dusty-turner/ggvfields/}
}
}
