% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{rad2deg}
\alias{rad2deg}
\title{Radians to Degrees Conversion}
\usage{
rad2deg(rad, rotate = 0)
}
\arguments{
\item{rad}{A numeric value representing an angle in radians.}

\item{rotate}{A numeric value to rotate the angle. Default is 0.}
}
\value{
The angle in degrees.
}
\description{
This function converts an angle from radians to degrees.
}
\keyword{internal}
