% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_vector.R
\name{scale_length_continuous}
\alias{scale_length_continuous}
\title{Create a Continuous Scale for Vector Length}
\usage{
scale_length_continuous(range = c(0.1, 0.5), ...)
}
\arguments{
\item{range}{rescales the input to the specific output range. Numeric vector of
length two, giving range of possible values. Should be between 0 and 1}

\item{...}{Other arguments passed to \code{continuous_scale()}.}
}
\description{
\code{scale_length_continuous} provides a continuous scale for controlling the length
aesthetic in a ggplot. This is particularly useful when working with vector plots
where vector lengths are mapped to a continuous scale.
}
