% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_vector_field.R
\name{stat_vector_field}
\alias{stat_vector_field}
\title{stat_vector_field}
\usage{
stat_vector_field(
  mapping = NULL,
  data = NULL,
  geom = "segment",
  position = "identity",
  ...,
  fun = NULL,
  scaling_factor = 1,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{inherit.aes}{}
}
\description{
stat_vector_field
}
