% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_vector_field.R
\docType{data}
\name{StatComplexVectorField}
\alias{StatComplexVectorField}
\title{Create a Complex Vector Field Plot Layer}
\format{
An object of class \code{StatComplexVectorField} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatComplexVectorField
}
\arguments{
\item{fun}{A user-defined function that takes a complex number as input and returns a complex number.}

\item{relim}{A numeric vector of length 2 giving the real-axis limits.}

\item{imlim}{A numeric vector of length 2 giving the imaginary-axis limits.}

\item{n}{An integer specifying the number of grid points along each axis.}
}
\value{
A ggplot2 layer that can be added to a ggplot object to produce a vector field plot.
}
\description{
\code{stat_complex_vector_field} generates a vector field plot layer using a user-defined function to compute the vector components. This is particularly useful for visualizing vector fields in a two-dimensional space.
}
\examples{
library(ggplot2)
# Example user-defined function
f <- function(z) (z^2 + 1) / (z^2 - 1)

# Create a ggplot with the vector field layer
ggplot() +
  geom_complex_vector_field(fun = f, relim = c(-2, 2), imlim = c(-2, 2), n = 100)

}
\keyword{datasets}
