% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{times}
\alias{times}
\title{Times Operator}
\usage{
times()
}
\arguments{
\item{x, }{y Numeric values to be multiplied.}
}
\value{
The product of x and y.
}
\description{
This function allows for multiplication with a custom operator.
}
\keyword{internal}
