% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_streamline.R
\docType{data}
\name{geom_streamplot}
\alias{geom_streamplot}
\alias{GeomStreamplot}
\alias{stat_streamplot}
\alias{StatStreamplot}
\title{Create a Stream Plot Geom Layer}
\usage{
geom_streamplot(
  mapping = NULL,
  data = NULL,
  stat = "streamplot",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = TRUE,
  inherit.aes = TRUE,
  fun,
  xlim = c(-1, 1),
  ylim = c(-1, 1),
  n = c(21, 21),
  mask_shape_type = "square",
  iterations = 100,
  stream_density = 1,
  method = "rk4",
  arrow = grid::arrow(angle = 20, length = unit(0.015, "npc"), type = "closed"),
  chop = TRUE,
  scale_stream = 1
)

stat_streamplot(
  mapping = NULL,
  data = NULL,
  stat = "streamplot",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = TRUE,
  inherit.aes = TRUE,
  fun,
  xlim = c(-1, 1),
  ylim = c(-1, 1),
  n = c(21, 21),
  mask_shape_type = "square",
  iterations = 100,
  stream_density = 1,
  method = "rk4",
  arrow = grid::arrow(angle = 20, length = unit(0.015, "npc"), type = "closed"),
  chop = TRUE,
  scale_stream = 1
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this layer.
When using a \verb{geom_*()} function to construct a layer, the \code{stat}
argument can be used the override the default coupling between geoms and
stats. The \code{stat} argument accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[ggplot2:layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[ggplot2:draw_key]{key glyphs}, to change the display of the layer in the legend.
}}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{fun}{A user-defined function that takes a numeric vector as input and
returns a numeric vector of the same length.}

\item{xlim}{A numeric vector of length 2 giving the x-axis limits.}

\item{ylim}{A numeric vector of length 2 giving the y-axis limits.}

\item{n}{A numeric vector of length 2 specifying the grid dimensions.}

\item{mask_shape_type}{A character string specifying the mask shape type:
"square", "diamond", "inset_square", or "circle".}

\item{iterations}{An integer specifying the number of steps to perform for
the numerical integration of the flow.}

\item{stream_density}{A numeric value that controls the density of the streamlines.
Higher values produce more streamlines. Default is 1.}

\item{method}{A character string specifying the ODE solver method to be used
(passed to \code{\link[deSolve:ode]{deSolve::ode()}}). Default is "rk4".}

\item{arrow}{Arrow specification, as created by \code{grid::arrow()}.}

\item{chop}{A logical value indicating whether to chop the trajectories into
segments.}

\item{scale_stream}{A numeric value specifying the maximum allowable segment
size as a fraction of the total range for chopping the trajectories.}
}
\value{
A ggplot2 layer that can be added to a ggplot object to produce a
stream plot.
}
\description{
\code{\link[=geom_streamplot]{geom_streamplot()}} generates a stream plot layer of a user-defined vector
field function. The lines in the plot represent the flow of data points
through the vector field.
}
\examples{

# Example user-defined function
f <- function(v) {
  x <- v[1]
  y <- v[2]
  c(-1 - x^2 + y, 1 + x - y^2)
}

ggplot() +
  geom_streamplot(
    fun = f, xlim = c(-3, 3), ylim = c(-3, 3), n = c(15, 15),
    iterations = 100, chop = TRUE, scale_stream = 1,
    mask_shape_type = "square", method = "rk4"
  ) +
  coord_fixed() +
  theme_minimal()
}
\keyword{datasets}
