% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_field.R
\docType{data}
\name{geom_vector_field}
\alias{geom_vector_field}
\alias{stat_vector_field}
\alias{StatVectorField}
\alias{GeomVectorField}
\title{Create a Vector Field Plot Layer}
\format{
An object of class \code{StatVectorField} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.

An object of class \code{GeomVectorField} (inherits from \code{GeomSegment}, \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
geom_vector_field(
  mapping = NULL,
  data = NULL,
  stat = "vectorfield",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  fun,
  xlim = c(-10, 10),
  ylim = c(-10, 10),
  v = c(1, 2),
  n = 16,
  center = TRUE,
  normalize = TRUE,
  scale_length = 1.5,
  arrow = grid::arrow(angle = 20, length = unit(0.015, "npc"), type = "closed"),
  ...
)

stat_vector_field(
  mapping = NULL,
  data = NULL,
  geom = "segment",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  fun,
  xlim = c(-10, 10),
  ylim = c(-10, 10),
  v = c(1, 2),
  n = 16,
  center = TRUE,
  normalize = TRUE,
  scale_length = 1.5,
  arrow = grid::arrow(angle = 20, length = unit(0.015, "npc"), type = "closed"),
  ...
)

StatVectorField

GeomVectorField
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this layer.
When using a \verb{geom_*()} function to construct a layer, the \code{stat}
argument can be used the override the default coupling between geoms and
stats. The \code{stat} argument accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{fun}{A user-defined function that takes two arguments (x and y
coordinates) and returns a list of two components: the x and y components
of the vector field.}

\item{xlim, ylim}{A numeric vector of length 2 giving the x/y-axis limits.}

\item{v}{Numeric vector specifying the direction vector components for
calculating the directional derivative.}

\item{n}{An integer specifying the number of grid points along each axis.}

\item{center}{Logical; if TRUE, centers the vector on the evaluated x/y
location. If FALSE, the vector origin is on the evaluated x/y location.}

\item{normalize}{Logical; if TRUE, normalizes the vector to a length of unit
1 and scales them to avoid overplotting based on grid density and plot
range.}

\item{scale_length}{Numeric; a scaling factor applied to the vectors to
adjust their length relative to the grid spacing. Defaults to 1.}

\item{arrow}{Arrow specification, as created by \code{grid::arrow()}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[ggplot2:layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[ggplot2:draw_key]{key glyphs}, to change the display of the layer in the legend.
}}

\item{geom}{The geometric object to use to display the data for this layer.
When using a \verb{stat_*()} function to construct a layer, the \code{geom} argument
can be used to override the default coupling between stats and geoms. The
\code{geom} argument accepts the following:
\itemize{
\item A \code{Geom} ggproto subclass, for example \code{GeomPoint}.
\item A string naming the geom. To give the geom as a string, strip the
function name of the \code{geom_} prefix. For example, to use \code{geom_point()},
give the geom as \code{"point"}.
\item For more information and other ways to specify the geom, see the
\link[ggplot2:layer_geoms]{layer geom} documentation.
}}
}
\value{
A ggplot2 layer that can be added to a ggplot object to produce a
vector field plot.
}
\description{
\code{geom_vector_field} generates a vector field plot layer using a
user-defined function to compute the vector components. This is particularly
useful for visualizing vector fields in a two-dimensional space.
}
\section{Computed variables}{


\describe{
\item{norm}{The magnitude of the vector field, computed as \eqn{\|\mathbf{w}\| = \sqrt{u^2 + v^2}}.}
\item{divergence}{The divergence of the vector field, computed as \eqn{\nabla \cdot \mathbf{F} = \frac{\partial f_x}{\partial x} + \frac{\partial f_y}{\partial y}}.}
\item{curl}{The curl of the vector field, computed as \eqn{\text{curl} \, \mathbf{F} = \frac{\partial f_y}{\partial x} - \frac{\partial f_x}{\partial y}}.}
\item{laplacian}{The Laplacian of the vector field, computed as \eqn{\Delta \mathbf{F} = \frac{\partial^2 f_x}{\partial x^2} + \frac{\partial^2 f_x}{\partial y^2} + \frac{\partial^2 f_y}{\partial x^2} + \frac{\partial^2 f_y}{\partial y^2}}.}
\item{directional_derivative}{The directional derivative of the vector field, computed as \eqn{D_{\mathbf{v}} \mathbf{F} = \frac{\partial F_x}{\partial x} v_x + \frac{\partial F_x}{\partial y} v_y + \frac{\partial F_y}{\partial x} v_x + \frac{\partial F_y}{\partial y} v_y}.}
}
}

\examples{

# example user-defined function
f <- function(v) {
  x <- v[1]; y <- v[2]
  c(-y, x)
}

# create a ggplot with the vector field layer
ggplot() +
  geom_vector_field(
    fun = f, xlim = c(-10, 10), ylim = c(-10, 10)
  )

# various options
ggplot() +
  geom_vector_field(
    aes(color = after_stat(norm)),
    fun = f, xlim = c(-10, 10), ylim = c(-10, 10),
    arrow = arrow(length = unit(1, "mm"))
  )

ggplot() +
  geom_vector_field(
    fun = f, xlim = c(-10, 10), ylim = c(-10, 10),
    n = 10, center = FALSE, normalize = FALSE,
    arrow = arrow(length = unit(1, "mm"))
  )

ggplot() +
  geom_vector_field(
    fun = f, xlim = c(-10, 10), ylim = c(-10, 10),
    center = TRUE, color = "black",
    scale_length = .7
  )

}
\keyword{datasets}
