% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_field.R
\docType{data}
\name{geom_vector_field}
\alias{geom_vector_field}
\alias{GeomVectorField}
\alias{stat_vector_field}
\alias{StatVectorField}
\title{Create a Vector Field Plot Layer}
\usage{
geom_vector_field(
  mapping = NULL,
  data = NULL,
  stat = "vectorfield",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  fun,
  xlim,
  ylim,
  n = 10,
  center = TRUE,
  normalize = TRUE,
  ...
)

stat_vector_field(
  mapping = NULL,
  data = NULL,
  geom = "segment",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  fun,
  xlim,
  ylim,
  n = 10,
  center = TRUE,
  normalize = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, either as a \code{ggproto} \code{Geom} subclass or as a string naming the
stat stripped of the \code{stat_} prefix (e.g. \code{"count"} rather than
\code{"stat_count"})}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{fun}{A user-defined function that takes two arguments (x and y
coordinates) and returns a list of two components: the x and y components
of the vector field.}

\item{xlim, ylim}{A numeric vector of length 2 giving the x-axis limits.}

\item{n}{An integer specifying the number of grid points along each axis.}

\item{center}{Logical; if TRUE, centers the vector on the evaluated x/y
location. If FALSE, the vector origin is on the evaluated x/y location.}

\item{normalize}{Logical; if TRUE, normalizes the vector to a length of unit
1.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{geom}{The geometric object to use to display the data, either as a
\code{ggproto} \code{Geom} subclass or as a string naming the geom stripped of the
\code{geom_} prefix (e.g. \code{"point"} rather than \code{"geom_point"})}
}
\value{
A ggplot2 layer that can be added to a ggplot object to produce a
vector field plot.
}
\description{
\code{\link[=geom_vector_field]{geom_vector_field()}} generates a vector field plot layer using a
user-defined function to compute the vector components. This is particularly
useful for visualizing vector fields in a two-dimensional space.
}
\examples{

# example user-defined function
f <- function(x, y) {
  u <- -y
  v <- x
  list(u, v)
}

# Create a ggplot with the vector field layer
ggplot() +
  geom_vector_field(
    fun = f, xlim = c(-10, 10), ylim = c(-10, 10), n = 20, center = TRUE,
    arrow = arrow(length = unit(1, "mm"))
  )

}
\keyword{datasets}
