% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_vector_field.R
\name{scale_length_continuous}
\alias{scale_length_continuous}
\title{Continuous Scale for Vector Length}
\usage{
scale_length_continuous(
  name = waiver(),
  n.breaks = waiver(),
  guide = "none",
  ...
)
}
\arguments{
\item{name}{The name of the scale.}

\item{n.breaks}{The number of breaks.}

\item{guide}{The guide to be used for this scale, such as a legend or color bar.
Can be a string specifying the type of guide, or a guide object returned by
a call to a guide function (e.g., \code{guide_legend} or \code{guide_colorbar}). Use \code{"none"}
to hide the guide.}

\item{...}{Additional parameters passed on to \code{continuous_scale}.}
}
\description{
\code{scale_length_continuous} provides a continuous scale for the length
aesthetic in \code{geom_vector_field}.
}
\examples{
f <- function(v) {
  x <- v[1]; y <- v[2]
  c(-y, x)
}
ggplot() +
  geom_vector_field(aes(length = after_stat(norm)), fun = f, xlim = c(-10, 10), ylim = c(-10, 10)) +
  scale_length_continuous()
}
