% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_vector.R
\name{scale_length_continuous}
\alias{scale_length_continuous}
\title{Create a Continuous Scale for Vector Length}
\usage{
scale_length_continuous(max_range = 0.5, ...)
}
\arguments{
\item{max_range}{The maximum value to which the input is rescaled. Numeric scalar
specifying the upper bound of the output range. Should be between 0 and 1.}

\item{...}{Other arguments passed to \code{continuous_scale()}.}
}
\description{
\code{scale_length_continuous} provides a continuous scale for controlling the length
aesthetic in a ggplot. This is particularly useful when working with vector plots
where vector lengths are mapped to a continuous scale.
}
